API: 1. Исправлен OrderController 2. Добавлено преобразование Enum в строки 3. Теперь имеется разделение для всех заказов и для конкретного заказа
Application:
1. Исправлен IOrdersService в методе GetAsync для возвращение OrderDetailsDTO
2. Добавлена модель OrderDetailsDTO

Domain:
1. Изменены модели и некоторые свойства

Infrastructure
1. Добавлена начальная миграция
2. Добавлены начальные данные для Order и OrderItems
3. Добавлены конфигурации для Order и OrderItems
4. Добавлен строитель для Order

Общее:
1. Исправлены опечатки

+++ RepositoryService/RepositoryService.Infrastructure/Persistence/Builders/ProductBuilder.cs
@@ -23,7 +23,7 @@ namespace RepositoryService.Infrastructure.Persistence.Builders
         /// <summary>
         /// Sets the product's name.
         /// </summary>
-        /// <param name="article">The article of the product.</param>
+        /// <param name="name">The name of the product.</param>
         /// <returns>The current <see cref="ProductBuilder"/> instance</returns>
         public ProductBuilder WithName(string name)
         {
@@ -58,7 +58,7 @@ namespace RepositoryService.Infrastructure.Persistence.Builders
         /// </summary>
         /// <param name="price">The price of the product.</param>
         /// <returns>The current <see cref="ProductBuilder"/> instance.</returns>
-        public ProductBuilder WithPrice(float price)
+        public ProductBuilder WithPrice(decimal price)
         {
             _product.Price = price;
             return this;
